changes Done

+++ src/main/java/com/demo/EMS/ServiceImpl/EmployeeServiceImpl.java
@@ -2,7 +2,9 @@ package com.demo.EMS.ServiceImpl;
 
 import com.demo.EMS.Repository.EmployeeRepo;
 import com.demo.EMS.Services.EmployeeService;
-import com.demo.EMS.beans.Employee;
+import com.demo.EMS.Entity.Employee;
+import com.demo.EMS.dto.AddEmployeeRequest;
+import com.demo.EMS.dto.UpdateEmployeeRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -22,22 +24,47 @@ public class EmployeeServiceImpl implements EmployeeService {
         return employeeRepo.findById(id).orElse(null) ;
     }
 
-    public void addEmployee(Employee e){
-        employeeRepo.save(e);
+    public Employee addEmployee(AddEmployeeRequest e){
+        Employee emp = new Employee(e.getFirstName(),e.getLastName(),e.getEmail());
+        return employeeRepo.save(emp);
 
     }
 
-    public void updateEmployee(Employee emp , int id){
-        if(id == emp.getId()) {
-            employeeRepo.save(emp);
+    public Employee updateEmployee(UpdateEmployeeRequest emp , int id){
+        Employee e = employeeRepo.findById(id).orElse(null);
+
+        if(e != null){
+            if(emp.getFirstName()!=null)
+                e.setFirstName(emp.getFirstName());
+            if(emp.getLastName()!=null)
+                e.setLastName(emp.getLastName());
+            if(emp.getEmail()!=null)
+                e.setEmail(emp.getEmail());
         }
+        employeeRepo.save(e);
+        return e;
+
     }
 
-    public void deleteAllEmployees(){
+    public int deleteAllEmployees(){
+        //employeeRepo.deleteAll();
+        int  count= (int) employeeRepo.count();
         employeeRepo.deleteAll();
+        return count;
     }
 
-    public void deleteEmployeeByID(int id){
-        employeeRepo.deleteById(id);
+    public boolean deleteEmployeeByID(int id){
+
+        //employeeRepo.deleteById(id);
+        Employee e = employeeRepo.findById(id).orElse(null);
+        if(e==null)
+        {
+
+            return false;
+        }
+        else{
+            employeeRepo.deleteById(id);
+            return true;
+        }
     }
 }
